2009-12-31  Sancho Lerena <slerena@artica.es>
        * pandoradb.sql: Custom graph weights now can be float.

        * include/config_process.php: Updated version to 3.1-dev

        This is the first commit of 3.1 version ! :-)



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@2261 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/pandoradb.sql
@@ -18,7 +18,7 @@
 -- Because Pandora Installer don't understand them
 -- and fails creating database !!!
 -- -----------------------------------------------------------
--- Pandora FMS official tables for 2.0 version              --
+-- Pandora FMS official tables for 3.1 version              --
 -- -----------------------------------------------------------
 
 CREATE TABLE IF NOT EXISTS `taddress` (
@@ -705,7 +705,7 @@ CREATE TABLE IF NOT EXISTS `tgraph_source` (
   `id_gs` INTEGER UNSIGNED NOT NULL  AUTO_INCREMENT,
   `id_graph` int(11) NOT NULL default 0,
   `id_agent_module` int(11) NOT NULL default 0,
-  `weight` smallint(5) UNSIGNED NOT NULL DEFAULT 0,
+  `weight` float(5,3) UNSIGNED NOT NULL DEFAULT 0,
   PRIMARY KEY(`id_gs`)
 ) ENGINE = InnoDB DEFAULT CHARSET=utf8;
